+++ lightkurve/correctors/pldcorrector.py
@@ -518,3 +518,4 @@ class TessPLDCorrector(PLDCorrector):
                                                     **kwargs)
 
         return clc
+