+++ lightkurve/tests/test_lightcurve.py
@@ -19,6 +19,8 @@ from ..lightcurve import LightCurve, KeplerLightCurve, TessLightCurve
 from ..lightcurvefile import KeplerLightCurveFile, TessLightCurveFile
 from ..targetpixelfile import KeplerTargetPixelFile, TessTargetPixelFile
 from ..utils import LightkurveWarning, LightkurveDeprecationWarning
+from ..search import search_lightcurve
+from ..collections import LightCurveCollection
 from .test_targetpixelfile import TABBY_TPF
 
 
@@ -122,7 +124,7 @@ def test_KeplerLightCurveFile(path, mission):
     assert lc.label == hdu[0].header['OBJECT']
     nanmask = ~np.isnan(hdu[1].data['TIME'])
     assert_array_equal(lc.time.value, hdu[1].data['TIME'][nanmask])
-    assert_array_equal(lc.flux.value, hdu[1].data['SAP_FLUX'][nanmask] / ((hdu[1].header['CROWDSAP'] * hdu[1].header['FLFRCSAP'])))
+    assert_array_equal(lc.flux.value, hdu[1].data['SAP_FLUX'][nanmask])
 
 
 @pytest.mark.remote_data
@@ -130,14 +132,13 @@ def test_KeplerLightCurveFile(path, mission):
                          ['hardest', 'hard', 'default', None,
                           1, 100, 2096639])
 def test_TessLightCurveFile(quality_bitmask):
-    tess_file = TessLightCurveFile.read(TESS_SIM, quality_bitmask=quality_bitmask)
+    lc = TessLightCurveFile.read(TESS_SIM, quality_bitmask=quality_bitmask, flux_column="sap_flux")
     hdu = pyfits.open(TESS_SIM)
-    lc = tess_file.SAP_FLUX
 
     assert lc.mission == 'TESS'
     assert lc.label == hdu[0].header['OBJECT']
-    assert lc.time_format == 'btjd'
-    assert lc.time_scale == 'tdb'
+    assert lc.time.format == 'btjd'
+    assert lc.time.scale == 'tdb'
     assert lc.flux.unit == u.electron / u.second
     assert lc.sector == hdu[0].header['SECTOR']
     assert lc.camera == hdu[0].header['CAMERA']
@@ -164,8 +165,8 @@ def test_bitmasking(quality_bitmask, answer):
 
 def test_lightcurve_fold():
     """Test the ``LightCurve.fold()`` method."""
-    lc = LightCurve(time=np.linspace(0, 10, 100), flux=np.zeros(100)+1,
-                    targetid=999, label='mystar', meta={'ccd': 2}, time_format='bkjd')
+    lc = KeplerLightCurve(time=np.linspace(0, 10, 100), flux=np.zeros(100)+1,
+                          targetid=999, label='mystar', meta={'ccd': 2})
     fold = lc.fold(period=1)
     assert_almost_equal(fold.phase[0], -0.5, 2)
     assert_almost_equal(np.min(fold.phase), -0.5, 2)
@@ -176,7 +177,7 @@ def test_lightcurve_fold():
     assert lc.meta['ccd'] == fold.meta['ccd']
     assert_array_equal(np.sort(fold.time_original), lc.time)
     assert len(fold.time_original) == len(lc.time)
-    fold = lc.fold(period=1, t0=-0.1)
+    fold = lc.fold(period=1, epoch_time=-0.1)
     assert_almost_equal(fold.time[0], -0.5, 2)
     assert_almost_equal(np.min(fold.phase), -0.5, 2)
     assert_almost_equal(np.max(fold.phase), 0.5, 2)
@@ -323,10 +324,10 @@ def test_custom_lightcurve_file(path, mission):
 
     # TESS has QUALITY while Kepler/K2 has SAP_QUALITY:
     if mission == "TESS":
-        assert "QUALITY" in lc.hdu[1].columns.names
+        assert "QUALITY" in hdu[1].columns.names
         assert_array_equal(lc.quality, hdu[1].data['QUALITY'])
     if mission in ["K2", "Kepler"]:
-        assert "SAP_QUALITY" in lc.hdu[1].columns.names
+        assert "SAP_QUALITY" in hdu[1].columns.names
         assert_array_equal(lc.quality, hdu[1].data['SAP_QUALITY'])
 
 
@@ -343,7 +344,11 @@ def test_lightcurve_plots():
         lc.scatter()
         lc.scatter(c='C3')
         lc.scatter(c=lc.time.value, show_colorbar=True, colorbar_label='Time')
-        lc.errorbar()
+        lc.plot(column='sap_flux')
+        lc.plot(column='sap_bkg', normalize=True)
+        lc.plot(column='cadenceno')
+        lc.errorbar(column='psf_centr1')
+        lc.errorbar(column='timecorr')
         plt.close('all')
 
 
@@ -625,9 +630,9 @@ def test_to_fits():
 
 
 def test_astropy_time_bkjd():
-    """Does `LightCurve.astropy_time` support bkjd?"""
+    """Does `KeplerLightCurve` support bkjd?"""
     bkjd = np.array([100, 200])
-    lc = LightCurve(time=[100, 200], time_format='bkjd')
+    lc = KeplerLightCurve(time=[100, 200])
     assert_allclose(lc.time.jd, bkjd + 2454833.)
 
 
@@ -839,20 +844,6 @@ def test_regression_346():
         KeplerLightCurveFile(K2_C08).PDCSAP_FLUX.remove_nans().to_corrector().correct().estimate_cdpp()
 
 
-def test_to_timeseries():
-    """Test the `LightCurve.to_timeseries()` method."""
-    time, flux, flux_err = np.arange(3)+2457576.4, np.ones(3), np.ones(3)*0.01
-    lc = LightCurve(time=time, flux=flux, flux_err=flux_err, time_format="jd")
-    try:
-        ts = lc.to_timeseries()
-        assert_allclose(ts['time'].value, time)
-        assert_allclose(ts['flux'], flux)
-        assert_allclose(ts['flux_err'], flux_err)
-    except ImportError:
-        # Requires AstroPy v3.2 or later
-        pass
-
-
 def test_flux_unit():
     """Checks the use of lc.flux_unit and lc.flux_quantity."""
     with warnings.catch_warnings():  # We deprecated `flux_unit` in v2.0
@@ -882,20 +873,20 @@ def test_flux_unit():
 def test_astropy_time_initialization():
     """Does the `LightCurve` constructor accept Astropy time objects?"""
     time = [1, 2, 3]
-    lc = LightCurve(time=Time(time, format='jd', scale='utc'))
-    assert lc.time_format == 'jd'
-    assert lc.time_scale == 'utc'
+    lc = LightCurve(time=Time(2.454e6+np.array(time), format='jd', scale='utc'))
+    assert lc.time.format == 'jd'
+    assert lc.time.scale == 'utc'
     with warnings.catch_warnings():  # we deprecated `astropy_time` in v2.0
         warnings.simplefilter("ignore", LightkurveDeprecationWarning)
         assert lc.astropy_time.format == 'jd'
         assert lc.astropy_time.scale == 'utc'
-    lc = LightCurve(time=time, time_format='jd', time_scale='utc')
-    assert lc.time_format == 'jd'
-    assert lc.time_scale == 'utc'
+    lc = LightCurve(time=time, time_format='bkjd', time_scale='tdb')
+    assert lc.time.format == 'bkjd'
+    assert lc.time.scale == 'tdb'
     with warnings.catch_warnings():  # we deprecated `astropy_time` in v2.0
         warnings.simplefilter("ignore", LightkurveDeprecationWarning)
-        assert lc.astropy_time.format == 'jd'
-        assert lc.astropy_time.scale == 'utc'
+        assert lc.astropy_time.format == 'bkjd'
+        assert lc.astropy_time.scale == 'tdb'
 
 
 def test_normalize_unit():
@@ -909,7 +900,7 @@ def test_normalize_unit():
 def test_to_stingray():
     """Test the `LightCurve.to_stingray()` method."""
     time, flux, flux_err = range(3), np.ones(3), np.zeros(3)
-    lc = LightCurve(time=time, flux=flux, flux_err=flux_err, time_format="jd")
+    lc = LightCurve(time=time, flux=flux, flux_err=flux_err)
     try:
         with warnings.catch_warnings():
             # Ignore "UserWarning: Numba not installed" raised by stingray.
@@ -938,7 +929,9 @@ def test_from_stingray():
 
 
 def test_river():
-    lc = LightCurve(time=np.arange(100), flux=np.random.normal(1, 0.01, 100))
+    lc = LightCurve(time=np.arange(100),
+                    flux=np.random.normal(1, 0.01, 100),
+                    flux_err=np.random.normal(0, 0.01, 100))
     lc.plot_river(10, 1)
     plt.close()
     folded_lc = lc.fold(10, 1)
@@ -1007,3 +1000,17 @@ def test_fold_v2():
     assert isinstance(fld.time, u.Quantity)
     fld.plot_river()
     plt.close()
+
+
+@pytest.mark.remote_data
+def test_combine_kepler_tess():
+    """Can we append or stitch a TESS light curve to a Kepler light curve?"""
+    lc_kplr = search_lightcurve("Kepler-10", mission='Kepler')[0].download()
+    lc_tess = search_lightcurve("Kepler-10", mission='TESS')[0].download()
+    # Can we use append()?
+    lc = lc_kplr.append(lc_tess)
+    assert(len(lc) == len(lc_kplr)+len(lc_tess))
+    # Can we use stitch()?
+    coll = LightCurveCollection((lc_kplr, lc_tess))
+    lc = coll.stitch()
+    assert(len(lc) == len(lc_kplr)+len(lc_tess))