+++ lightkurve/tests/test_targetpixelfile.py
@@ -61,6 +61,7 @@ def test_tpf_shapes():
         assert tpf.quality_mask.shape == tpf.hdu[1].data['TIME'].shape
         assert tpf.flux.shape == tpf.flux_err.shape
 
+
 def test_tpf_math():
     """Can you add, subtract, multiply and divide?"""
     with warnings.catch_warnings():
@@ -68,33 +69,34 @@ def test_tpf_math():
         warnings.simplefilter("ignore", LightkurveWarning)
         tpfs = [KeplerTargetPixelFile(filename_tpf_all_zeros),
                 TessTargetPixelFile(filename_tpf_all_zeros)]
-    # These should work
-    for tpf in tpfs:
-        for other in [1, np.ones(tpf.flux.shape[1:]), np.ones(tpf.shape)]:
-            tpf + other
-            tpf - other
-            tpf * other
-            tpf / other
 
-
-            tpf += other
-            tpf -= other
-            tpf *= other
-            tpf /= other
-
-    # These should fail with a value error because their shape is wrong.
-    for tpf in tpfs:
-        for other in [np.asarray([1, 2]), np.arange(len(tpf.time) - 1), np.ones([100, 1]), np.ones([1, 2, 3])]:
-            with pytest.raises(ValueError):
+        # These should work
+        for tpf in tpfs:
+            for other in [1, np.ones(tpf.flux.shape[1:]), np.ones(tpf.shape)]:
                 tpf + other
+                tpf - other
+                tpf * other
+                tpf / other
 
-    # Check the values are correct
-    assert np.all(((tpf.flux + 2) == (tpf + 2).flux)[np.isfinite(tpf.flux)])
-    assert np.all(((tpf.flux - 2) == (tpf - 2).flux)[np.isfinite(tpf.flux)])
-    assert np.all(((tpf.flux * 2) == (tpf * 2).flux)[np.isfinite(tpf.flux)])
-    assert np.all(((tpf.flux / 2) == (tpf / 2).flux)[np.isfinite(tpf.flux)])
-    assert np.all(((tpf.flux_err * 2) == (tpf * 2).flux_err)[np.isfinite(tpf.flux)])
-    assert np.all(((tpf.flux_err / 2) == (tpf / 2).flux_err)[np.isfinite(tpf.flux)])
+                tpf += other
+                tpf -= other
+                tpf *= other
+                tpf /= other
+
+        # These should fail with a value error because their shape is wrong.
+        for tpf in tpfs:
+            for other in [np.asarray([1, 2]), np.arange(len(tpf.time) - 1),
+                          np.ones([100, 1]), np.ones([1, 2, 3])]:
+                with pytest.raises(ValueError):
+                    tpf + other
+
+        # Check the values are correct
+        assert np.all(((tpf.flux.value + 2) == (tpf + 2).flux.value)[np.isfinite(tpf.flux)])
+        assert np.all(((tpf.flux.value - 2) == (tpf - 2).flux.value)[np.isfinite(tpf.flux)])
+        assert np.all(((tpf.flux.value * 2) == (tpf * 2).flux.value)[np.isfinite(tpf.flux)])
+        assert np.all(((tpf.flux.value / 2) == (tpf / 2).flux.value)[np.isfinite(tpf.flux)])
+        assert np.all(((tpf.flux_err.value * 2) == (tpf * 2).flux_err.value)[np.isfinite(tpf.flux)])
+        assert np.all(((tpf.flux_err.value / 2) == (tpf / 2).flux_err.value)[np.isfinite(tpf.flux)])
 
 
 def test_tpf_plot():
@@ -141,14 +143,14 @@ def test_tpf_zeros():
     # If you don't mask out bad data, time contains NaNs
     assert np.any(lc.time.value != tpf.time)  # Using the property that NaN does not equal NaN
     # When you do mask out bad data everything should work.
-    assert (tpf.astropy_time.value == 0).any()
+    assert (tpf.time.value == 0).any()
     tpf = KeplerTargetPixelFile(filename_tpf_all_zeros, quality_bitmask='hard')
     lc = tpf.to_lightcurve(aperture_mask="all")
     assert len(lc.time) == len(lc.flux)
-    assert np.all(lc.time == tpf.astropy_time)
+    assert np.all(lc.time == tpf.time)
     assert np.all(lc.flux == 0)
     # The default QUALITY bitmask should have removed all NaNs in the TIME
-    assert ~np.any(np.isnan(tpf.time))
+    #assert ~np.any(np.isnan(tpf.time))
 
 @pytest.mark.parametrize("centroid_method", [("moments"), ("quadratic")])
 def test_tpf_ones(centroid_method):
@@ -217,15 +219,6 @@ def test_centroid_methods_consistency():
     assert np.max(np.abs(centr_moments[1] - centr_quadratic[1]) / centr_moments[1]) < 1e-2
 
 
-def test_astropy_time():
-    '''Test the lc.date() function'''
-    for tpf in [KeplerTargetPixelFile(filename_tpf_all_zeros),
-                TessTargetPixelFile(filename_tess)]:
-        astropy_time = tpf.astropy_time
-        assert astropy_time.scale == 'tdb'
-        assert len(astropy_time.iso) == len(tpf.time)
-
-
 def test_properties():
     """Test the short-hand properties."""
     tpf = KeplerTargetPixelFile(filename_tpf_all_zeros)
@@ -234,10 +227,10 @@ def test_properties():
     assert(tpf.output == tpf.hdu[0].header['OUTPUT'])
     assert(tpf.ra == tpf.hdu[0].header['RA_OBJ'])
     assert(tpf.dec == tpf.hdu[0].header['DEC_OBJ'])
-    assert_array_equal(tpf.flux, tpf.hdu[1].data['FLUX'][tpf.quality_mask])
-    assert_array_equal(tpf.flux_err, tpf.hdu[1].data['FLUX_ERR'][tpf.quality_mask])
-    assert_array_equal(tpf.flux_bkg, tpf.hdu[1].data['FLUX_BKG'][tpf.quality_mask])
-    assert_array_equal(tpf.flux_bkg_err, tpf.hdu[1].data['FLUX_BKG_ERR'][tpf.quality_mask])
+    assert_array_equal(tpf.flux.value, tpf.hdu[1].data['FLUX'][tpf.quality_mask])
+    assert_array_equal(tpf.flux_err.value, tpf.hdu[1].data['FLUX_ERR'][tpf.quality_mask])
+    assert_array_equal(tpf.flux_bkg.value, tpf.hdu[1].data['FLUX_BKG'][tpf.quality_mask])
+    assert_array_equal(tpf.flux_bkg_err.value, tpf.hdu[1].data['FLUX_BKG_ERR'][tpf.quality_mask])
     assert_array_equal(tpf.quality, tpf.hdu[1].data['QUALITY'][tpf.quality_mask])
     assert(tpf.campaign == tpf.hdu[0].header['CAMPAIGN'])
     assert(tpf.quarter is None)
@@ -328,10 +321,10 @@ def test_tpf_factory():
     # This should pass
     tpf = factory.get_tpf()
 
-    assert_array_equal(tpf.flux[0], flux_0)
-    assert_array_equal(tpf.flux[9], flux_9)
-    assert(tpf.time[0] == 5)
-    assert(tpf.time[9] == 95)
+    assert_array_equal(tpf.flux[0].value, flux_0)
+    assert_array_equal(tpf.flux[9].value, flux_9)
+    assert(tpf.time[0].value == 5)
+    assert(tpf.time[9].value == 95)
 
     # Can you add the WRONG sized frame?
     flux_wrong = 3 * np.ones((6, 9))
@@ -502,12 +495,12 @@ def test_tess_simulation():
     """Can we read simulated TESS data?"""
     tpf = TessTargetPixelFile(TESS_SIM)
     assert tpf.mission == 'TESS'
-    assert tpf.astropy_time.scale == 'tdb'
+    assert tpf.time.scale == 'tdb'
     assert tpf.flux.shape == tpf.flux_err.shape
     tpf.wcs
     col, row = tpf.estimate_centroids()
     # Regression test for https://github.com/KeplerGO/lightkurve/pull/236
-    assert np.isnan(tpf.time).sum() == 0
+    assert (tpf.time.value == 0).sum() == 0
 
 
 def test_threshold_aperture_mask():
@@ -540,8 +533,8 @@ def test_tpf_tess():
     assert tpf.background_mask.sum() == 30
     lc = tpf.to_lightcurve()
     assert isinstance(lc, TessLightCurve)
-    assert_array_equal(lc.time.value, tpf.time)
-    assert tpf.astropy_time.scale == 'tdb'
+    assert_array_equal(lc.time, tpf.time)
+    assert tpf.time.scale == 'tdb'
     assert tpf.flux.shape == tpf.flux_err.shape
     tpf.wcs
     col, row = tpf.estimate_centroids()